+++ src/ClientPage.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { LightboxClient } from './lib/core/index';
 import './lib/styles/styles.css';
 
@@ -17,15 +16,18 @@ const ClientPage = () => {
                     LightboxClient.init(`${window.location.origin}/src/appPage.html`, {
                         dispatch: true,
                         callbacks: {
-                            onClose: (data) => {
+                            close: (data) => {
                                 console.log('cerrado', data);
                             },
                         },
                         styles: {
                             background: {
-                                color: '#ff0',
+                                opacity: 1,
                             },
                             rounded: 16,
+                            height: 500,
+                            width: 800,
+                            dimension: 'lg',
                         },
                     })
                 }