+++ src/lib/core/LightboxClient/assemblers.ts
@@ -1,24 +1,37 @@
-import { Styles, Ids } from '@/constants';
+import { Config } from '@/config';
+import { Styles, ElementIds, Dimensions, LightboxAppEvents } from '@/constants';
 import { ClientCallbacks, ApiStructure, ClientStyles } from '@/types';
 
-export const mountListener = (callbacks: ClientCallbacks) => {
+export const mountListeners = (callbacks: ClientCallbacks, styles: ClientStyles) => {
     if (!callbacks) return;
 
-    const listener = (event: MessageEvent<ApiStructure>) => {
+    const callbacksListener = (event: MessageEvent<ApiStructure>) => {
+        if (event.data.event !== LightboxAppEvents.EMIT) return;
         callbacks[event.data.type]?.(event.data.data);
-        unmountLightbox(listener);
+        unmountLightbox(callbacksListener);
     };
 
-    globalThis.addEventListener('message', listener);
+    const stylesListener = (event: MessageEvent<{ event: LightboxAppEvents; styles: ClientStyles }>) => {
+        if (event.data.event !== LightboxAppEvents.SEND_STYLES) return;
+        mountStyles({
+            ...event.data.styles,
+            ...styles,
+            background: { ...event.data.styles.background, ...styles.background },
+        });
+        globalThis.removeEventListener('message', stylesListener);
+    };
+
+    globalThis.addEventListener('message', callbacksListener);
+    globalThis.addEventListener('message', stylesListener);
 };
 
 export const mountIFrameElement = (url: string, styles: ClientStyles) => {
     const wrapper = document.createElement('div');
-    wrapper.id = Ids.WRAPPER_ID;
+    wrapper.id = ElementIds.WRAPPER_ID;
 
     const iframe = document.createElement('iframe');
     iframe.src = url;
-    iframe.id = Ids.IFRAME_ID;
+    iframe.id = ElementIds.IFRAME_ID;
 
     mountStyles(styles);
 
@@ -34,16 +47,22 @@ const mountStyles = (styles: ClientStyles) => {
         ?.map((x) => parseInt(x, 16)) ?? [107, 114, 128];
 
     background.push(styles.background?.opacity ?? 0.75);
-
     document.documentElement.style.setProperty(Styles.BACKGROUND_COLOR, `rgb(${background.join(', ')})`);
 
     const rounded = styles.rounded ?? 0;
-
     document.documentElement.style.setProperty(Styles.ROUNDED, `${rounded.toString()}px`);
+
+    let height = styles.height ?? Config.defaultDimension.height;
+    if (styles.dimension) height = Dimensions[styles.dimension.toUpperCase() as keyof typeof Dimensions].height;
+    document.documentElement.style.setProperty(Styles.MAX_HEIGHT, `${height.toString()}px`);
+
+    let width = styles.width ?? Config.defaultDimension.width;
+    if (styles.dimension) width = Dimensions[styles.dimension.toUpperCase() as keyof typeof Dimensions].width;
+    document.documentElement.style.setProperty(Styles.MAX_WIDTH, `${width.toString()}px`);
 };
 
 const unmountLightbox = (listener: (event: MessageEvent<ApiStructure>) => void) => {
-    document.getElementById(Ids.WRAPPER_ID)?.remove();
+    document.getElementById(ElementIds.WRAPPER_ID)?.remove();
     globalThis.removeEventListener('message', listener);
     document.documentElement.style.removeProperty(Styles.BACKGROUND_COLOR);
 };