+++ src/lib/core/LightboxClient/index.ts
@@ -1,13 +1,16 @@
+import { redirectBasedOnDriver } from '@/helpers';
 import { InitOptions, LightboxInstance } from '@/types';
-import { mountIFrameElement, mountListener } from './assemblers';
+import { mountIFrameElement, mountListeners } from './assemblers';
 
 export const LightboxClient = {
     init: (url: string, options?: InitOptions): LightboxInstance => {
         const lightbox: LightboxInstance = {
             callbacks: options?.callbacks ?? {},
+            allowRedirects: options?.allowRedirects ?? true,
             styles: options?.styles ?? {},
             open: () => {
-                mountListener(lightbox.callbacks);
+                if (lightbox.allowRedirects) redirectBasedOnDriver(url);
+                mountListeners(lightbox.callbacks, lightbox.styles);
                 mountIFrameElement(url, lightbox.styles);
             },
         };